Added Display Functionality to Frontend
Now data is displayed properly and rounded and all. A Selector still needs to be added, auto stuff needs to be figured out still, and pick list needs to be written, but overall the frontend API integrations are working really well.

Something I feel like I still need to resolve is that right now the frontend calls the API pretty much any time data is updated.
The frequency of queries would optimally be lessened.